Update the line parsing for `airport` output
On macOS Monterey, the `airport -s` output doesn't contain the BSSID
unless you run it as root.  This causes the parsing to fail because
none of the lines match the BSSID regexp.  Instead, we update the
parsing to be simple column chopping because it's a nice fixed format.

Tested on Monterey and pre-Monterey output to verify that they're both
compatible with this approach.

(`airport`, at least on Monterey, does have an XML output option but
that's a whole other can of worms.)